Refactoring: PushContext, PopContext -> using RenderContext.
